+++ crates/tracel-core/src/experiment/remote/cloud/mod.rs
@@ -20,8 +20,9 @@ use tracel_experiment::ArtifactKind;
 use tracel_experiment::error::{ExperimentError, ExperimentErrorKind};
 use tracel_experiment::{CancelToken, ExperimentId, ExperimentRun};
 
+use tracel_experiment::ExperimentProvider;
+
 use crate::backend::cloud::CloudBackend;
-use crate::experiment::ExperimentProvider;
 use crate::experiment::remote::session::RemoteExperimentSession;
 
 #[derive(Debug, Clone)]
